refactor: separate Emoji to own module
